Dodałem klasę z testami algorytmów kompresji. Reszta roboty w ręce Adasia :)
